requirements.txt
aiohttp==3.7.4.post0
discord.py==1.7.1
dnspython==2.1.0
Pillow==8.2.0
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.17.0
requests==2.25.1
cppimport==21.3.7
pybind11==2.6.2
typing==3.7.4.3